training/requirements.txt
kagglehub>=0.3.0
ultralytics>=8.0.0
PyYAML>=6.0.0
opencv-python>=4.8.0
torch>=2.0.0
numpy>=1.24.0
